rename: AuditIcon.tsx→MonitoringIcon.tsx
