+++ components/side-bar/menu-item-icons/MonitoringIcon.tsx
