+++ components/side-bar/menu-item-icons/index.ts
@@ -4,5 +4,5 @@ export { OrderIcon } from "./OrderIcon";
 export { SalesIcon } from "./SalesIcon";
 export { ShiftIcon } from "./ShiftIcon";
 export { UserIcon } from "./UserIcon";
-export { AuditIcon } from "./AuditIcon";
+export { AuditIcon } from "./MonitoringIcon";
 export { SettingsIcon } from "./SettingsIcon";
